[Project setup]:     - project structure change     - entities primary keys type change to Long     - RequestDetails contains timestamp of request     - app has different context path now     - app speed significantly increased by extracting UserAgentParser to separate bean which now loads on startup (startup time of app bit decreased)     - introduction of ApiError object which serves as consistent error message response     - introduction of ApiExceptionHandler class which serves as centralized place for error handling in the app     - introduction of RequestsInspectorController for the purpose of REST API for the frontend part of the app     - app has DTO classes for the purpose of implementing HATEOAS principles     - implementation of basic HATEOAS logic and creation of "processor" classes for the purpose of creating HATEOAS links in DTO objects     - started implementation of frontend part of the app     - introduction of linter and prettier for the purpose of consistency of code style

+++ src/main/java/com/stek/webdash/service/EndpointMappingService.java
@@ -3,7 +3,7 @@ package com.stek.webdash.service;
 import java.util.Optional;
 import java.util.Set;
 
-import com.stek.webdash.model.EndpointMapping;
+import com.stek.webdash.model.domain.EndpointMapping;
 
 public interface EndpointMappingService {
 
@@ -11,7 +11,7 @@ public interface EndpointMappingService {
 
 	Set<EndpointMapping> findAllEndpointMappings();
 
-	Optional<EndpointMapping> findEndpointMappingById(Integer id);
+	Optional<EndpointMapping> findEndpointMappingById(Long id);
 
 	void deleteEndpointMapping(EndpointMapping endpointMapping);
