+++ src/main/java/com/stek/webdash/service/repository/EndpointMappingRepository.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.stek.webdash.model.EndpointMapping;
+import com.stek.webdash.model.domain.EndpointMapping;
 
 @Repository
-public interface EndpointMappingRepository extends JpaRepository<EndpointMapping, Integer> {
+public interface EndpointMappingRepository extends JpaRepository<EndpointMapping, Long> {
 
 	Optional<EndpointMapping> findByHttpMethodAndRequestUri(String httpMethod, String requestUri);
