+++ src/main/java/com/stek/webdash/service/repository/RequestDetailsRepository.java
@@ -3,8 +3,8 @@ package com.stek.webdash.service.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.stek.webdash.model.RequestDetails;
+import com.stek.webdash.model.domain.RequestDetails;
 
 @Repository
-public interface RequestDetailsRepository extends JpaRepository<RequestDetails, Integer> {
+public interface RequestDetailsRepository extends JpaRepository<RequestDetails, Long> {
 }
